+++ packages/site-demo/content/product/components/image-block/react/image-block-alignment.tsx
@@ -18,9 +18,7 @@ export const App = ({ theme }: any) => (
         }
         theme={theme}
         title="Lorem ipsum"
-        thumbnailProps={{
-            align: Alignment.center,
-            aspectRatio: AspectRatio.original,
-        }}
+        align={Alignment.center}
+        thumbnailProps={{ aspectRatio: AspectRatio.original }}
     />
 );